Refactor metric module : support TCP exporter and lineprotocol

+++ agent-plugins/agent-app-insight-common/src/main/java/cloud/erda/agent/plugin/app/insight/transaction/TransactionMetricBuilder.java
@@ -91,7 +91,7 @@ public class TransactionMetricBuilder implements MetricBuilder {
         return metric;
     }
 
-    private void recordElapsed(long elapsed) {
+    private void recordElapsed(float elapsed) {
         fields.put(Constants.Metrics.ELAPSED, elapsed);
     }
